Reorganize common data messages in Rust
