+++ crates/common/src/messages/execution/mod.rs
@@ -12,20 +12,3 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
-
-use nautilus_model::data::Data;
-
-pub mod data;
-pub mod execution;
-pub mod system;
-
-// Re-exports
-pub use data::{DataResponse, SubscribeCommand, UnsubscribeCommand};
-
-// TODO: Refine this to reduce disparity between enum sizes
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug)]
-pub enum DataEvent {
-    Response(DataResponse),
-    Data(Data),
-}
