+++ crates/common/src/messages/system/mod.rs
@@ -13,19 +13,7 @@
 //  limitations under the License.
 // -------------------------------------------------------------------------------------------------
 
-use nautilus_model::data::Data;
-
-pub mod data;
-pub mod execution;
-pub mod system;
+pub mod shutdown;
 
 // Re-exports
-pub use data::{DataResponse, SubscribeCommand, UnsubscribeCommand};
-
-// TODO: Refine this to reduce disparity between enum sizes
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug)]
-pub enum DataEvent {
-    Response(DataResponse),
-    Data(Data),
-}
+pub use shutdown::ShutdownSystem;
